+++ tools/pylint-extensions/azure-pylint-guidelines-checker/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add `remove-deprecated-iscoroutinefunction` check
 - Add `do-not-use-logging-directly` check
 - Add `no-cross-package-private-import` check
+- Fix `check-docstrings` to skip *args/**kwargs linting for overload implementations
 
 ## 0.5.7 (2025-07-15)
 - Bug fix for `do-not-use-logging-exception` checker